* commit 'b4862141a68b32cc0b5790819d0789c2fbbad1d3':
  cd_pbb: Fix format string warning for ethertype

+++ src/codecs/cd_pbb/cd_pbb.cc
@@ -122,7 +122,7 @@ void PbbCodec::log(TextLog* const text_log, const uint8_t* raw_pkt,
         pbb->ether_dst[4], pbb->ether_dst[5]);
 
     TextLog_Print(text_log, "  len:%hu", len);
-    TextLog_Print(text_log, "  type:0x%04X", pbb->ethertype());
+    TextLog_Print(text_log, "  type:0x%04X", static_cast<uint16_t>(pbb->ethertype()));
 }
 
 //-------------------------------------------------------------------------
